Rewrote / Added Files to MVP
I am trying to move forward with a minimum viable product, to showcase a lot of our work. In this, I rewrote a lot of the code that we previously had / copy pasting. Not because it was not functional (it was), but I just cannot function after so long without knowing what every piece of the code does.